pyproject.toml
[tool.poetry]
name = "dp-cgans"
version = "0.0.4"
description = "A library to generate synthetic tabular data using Conditional Generative Adversary Networks (GANs) combined with Differential Privacy techniques."
# description = "This package is extended from SDV (https://github.com/sdv-dev/SDV), CTGAN (https://github.com/sdv-dev/CTGAN), and Differential Privacy in GANs (https://github.com/civisanalytics/dpwgan). The author modified the conditional matrix and cost functions to emphasize on the relations between variables. The main changes are in ctgan/synthesizers/ctgan.py ../data_sampler.py ../data_transformer.py"
authors = ["Chang Sun <sunchang0124@gmail.com>", "Leopold Cudilla <leopoldcudilla@gmail.com>", "Vincent Emonet <vincent.emonet@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/sunchang0124/dp_cgans"
repository = "https://github.com/sunchang0124/dp_cgans"
keywords = [ "CGAN", "synthetic data", "DP", "Differential Privacy", "GAN" ]
include = [
    "LICENSE",
]
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]


[tool.poetry.scripts]
dp-cgans = "dp_cgans.__main__:cli"


[tool.poetry.dependencies]
python = ">=3.8,<3.10"
wheel = "*"
numpy = "*"
pandas = "*"
torch = "*"
scipy = "*"
rdt= "0.6.4"
scikit-learn = "*"
graphviz = "*"
faker = "*"
copulas = "*"
pyreadstat = "*"
sdv = "0.14.0"
typer = "*"
mowl-borg = {path = "mowl/mowl_borg-0.2.0-py38-none-any.whl", develop = true}
owlready2 = "^0.41"
rdflib = "^6.3.2"
jupyter = "^1.0.0"
nltk = "^3.8.1"
owl2vec-star = {git = "https://github.com/cudillal/OWL2Vec-Star.git"}
oxrdflib = "^0.3.4"
transformers = "^4.29.2"
sentence-transformers = "^2.2.2"
progress = "^1.6"


[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"


[tool.poetry.urls]
"Bug Tracker" = "https://github.com/sunchang0124/dp_cgans/issues"


[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
